Mass Refact(or)
refactored to use channels better, for communication relaying and
created the "switch board" to effectively handle this. Renamed some
the "printing" text to be a bit more device neutral by using "job"
instead. Included a jIn channel to serve as the job queue and will
log the commands over to the device during an "extended job run" such
as a print