Index Out of Range
Fixed an ugly index out of range bug when running an extended job (such
as a print file. Also updated the main css file to have the x/y labels
on a lower z-index (was 900, now 30) so that they're still on top of the
controller, but not on top of the fixed "nav" bar. Small tweaks to the
.js to allow the user to "force" a checkConn by clicking on the screen
during an "init"